requirements.txt
# FastAPI и веб-фреймворк
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# Асинхронные HTTP клиенты
aiohttp==3.9.1
httpx==0.25.2
aiofiles==23.2.1

# Парсинг HTML
beautifulsoup4==4.12.2
lxml==4.9.3

# База данных
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Кэширование и очереди
redis>=4.5.2,<5.0.0
aioredis==2.0.1
aiosqlite==0.19.0
celery[redis]==5.3.4

# Внешние сервисы
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Telegram Bot
python-telegram-bot==20.7
pydantic==2.5.0

# Утилиты
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
requests==2.31.0
psutil==5.9.6

# Логирование и мониторинг
structlog==23.2.0
prometheus-client==0.19.0

# Разработка и тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
